requirements.txt
requests
pandas
pyarrow
sqlalchemy
pymysql
python-dotenv
tenacity
cryptography
